+++ .github/component_owners.yml
@@ -5,8 +5,6 @@ components:
   src/OpenTelemetry.Contrib.Extensions.AWSXRay/:
     - srprash
     - lupengamzn
-  src/OpenTelemetry.Contrib.Instrumentation.GrpcCore/:
-    - pcwiese
   src/OpenTelemetry.Contrib.Instrumentation.MassTransit/:
     - alexvaluyskiy
   src/OpenTelemetry.Contrib.Instrumentation.Owin/:
@@ -21,6 +19,8 @@ components:
     - SergeyKanzhelev
   src/OpenTelemetry.Instrumentation.ElasticsearchClient/:
     - ejsmith
+  src/OpenTelemetry.Instrumentation.GrpcCore/:
+    - pcwiese
   src/OpenTelemetry.Instrumentation.Quartz/:
     - maldago
   src/OpenTelemetry.Instrumentation.Wcf/:
@@ -29,8 +29,6 @@ components:
   test/OpenTelemetry.Contrib.Extensions.AWSXRay.Tests/:
     - srprash 
     - lupengamzn
-  test/OpenTelemetry.Contrib.Instrumentation.GrpcCore.Tests/:
-    - pcwiese
   test/OpenTelemetry.Contrib.Instrumentation.MassTransit.Tests/:
     - alexvaluyskiy
   test/OpenTelemetry.Contrib.Instrumentation.Owin.Tests/:
@@ -45,6 +43,8 @@ components:
     - SergeyKanzhelev
   test/OpenTelemetry.Instrumentation.ElasticsearchClient.Tests/:
     - ejsmith
+  test/OpenTelemetry.Instrumentation.GrpcCore.Tests/:
+    - pcwiese
   test/OpenTelemetry.Instrumentation.Quartz.Tests/:
     - maldago
   test/OpenTelemetry.Instrumentation.Wcf.Tests/:
