+++ examples/grpc.core/Examples.GrpcCore.AspNetCore/Examples.GrpcCore.AspNetCore.csproj
@@ -17,7 +17,7 @@
     <Protobuf Include="proto/*.proto" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="$(RepoRoot)\src\OpenTelemetry.Contrib.Instrumentation.GrpcCore\OpenTelemetry.Instrumentation.GrpcCore.csproj" />
+    <ProjectReference Include="$(RepoRoot)\src\OpenTelemetry.Instrumentation.GrpcCore\OpenTelemetry.Instrumentation.GrpcCore.csproj" />
   </ItemGroup>
 
 </Project>
