--- src/OpenTelemetry.Contrib.Instrumentation.GrpcCore/CHANGELOG.md
@@ -1,9 +1,0 @@
-# Changelog
-
-## Unreleased
-
-* Updated OTel SDK package version to 1.1.0-beta1
-  ([#100](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/100))
-
-* Do NOT mutate incoming call headers, copy them before propagation
-  ([#143](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/143))
