+++ src/OpenTelemetry.Instrumentation.GrpcCore/AssemblyInfo.cs
