+++ test/OpenTelemetry.Instrumentation.GrpcCore.Tests/OpenTelemetry.Instrumentation.GrpcCore.Tests.csproj
@@ -20,6 +20,6 @@
     <Protobuf Include="proto/*.proto" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="$(RepoRoot)\src\OpenTelemetry.Contrib.Instrumentation.GrpcCore\OpenTelemetry.Instrumentation.GrpcCore.csproj" />
+    <ProjectReference Include="$(RepoRoot)\src\OpenTelemetry.Instrumentation.GrpcCore\OpenTelemetry.Instrumentation.GrpcCore.csproj" />
   </ItemGroup>
 </Project>
